refactor: some adjusts at tab bar class
