+++ solutions/devsprint-julio-fernandes-6/FinanceApp/AppDelegate/SceneDelegate.swift
@@ -13,11 +13,15 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
 
 
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        
+        if #available(iOS 15, *) {
+            UINavigationBar.appearance().scrollEdgeAppearance = UINavigationBarAppearance()
+        }
 
         guard let windowScene = (scene as? UIWindowScene) else { return }
 
         self.window = UIWindow(frame: UIScreen.main.bounds)
-        self.window?.rootViewController = UINavigationController(rootViewController: MainViewController())
+        self.window?.rootViewController = UINavigationController(rootViewController: TabbarController())
         self.window?.windowScene = windowScene
         self.window?.makeKeyAndVisible()
     }
